hollaex-cli rebranding to hex-cli. - v1.5.0

+++ uninstall.sh
@@ -2,9 +2,9 @@
 
 REPLACE_EXISTING_TO_LATEST=false
 
-echo "#### hollaex-cli Uninstaller ####"
+echo "#### hex-cli Uninstaller ####"
 
-echo "Are you sure you want to uninstall hollaex-cli from your computer? (y/n)"
+echo "Are you sure you want to uninstall hex-cli from your computer? (y/n)"
 
 read answer
 
@@ -13,8 +13,14 @@ if [ "$answer" != "${answer#[Nn]}" ] ;then
     exit 0;
 fi
 
-sudo rm /usr/local/bin/hollaex
-sudo rm -r $HOME/.hollaex-cli
+# Remove old hollaex-cli related files together If it's left on system.
+if [[ -d "$HOME/.hollaex-cli" ]]; then
+    sudo rm /usr/local/bin/hollaex
+    sudo rm -r $HOME/.hollaex-cli
+fi
+
+sudo rm /usr/local/bin/hex
+sudo rm -r $HOME/.hex-cli
 
 echo "Jobs all done!"
-echo "If you want to reinstall hollaex-cli later, Please visit https://github.com/bitholla/hollaex-cli for further information."
+echo "If you want to reinstall hex-cli later, Please visit https://github.com/bitholla/hex-cli for further information."
